requirements.txt
aiosmtpd                    # debian: python3-aiosmtpd
bleach                      # debian: python3-bleach
django                      # debian: python3-django
django-allauth              # debian (buster): python3-django-allauth
django-crispy-forms         # debian: python3-django-crispy-forms
django-filter               # debian: python3-django-filters
django-haystack             # debian: python3-django-haystack
django-ical                 # NOT IN DEBIAN
django-imagekit             # debian (buster): python3-django-imagekit
django-mailbox              # NOT IN DEBIAN
django-mailer               # NOT IN DEBIAN
djangorestframework         # debian: python-djangorestframework
feedparser                  # debian: python3-feedparser
flake8                      # debian: flake8
html2text                   # debian: python3-html2text
markdown                    # debian: python3-markdown
mdx_unimoji                 # NOT IN DEBIAN
pillow                      # debian: python-pil
pymdown-extensions          # NOT IN DEBIAN
aiosmtplib                  # NOT IN DEBIAN
schulze                     # NOT IN DEBIAN
randomcolor                 # NOT IN DEBIAN
raven                       # debian (buster): python3-raven
requests                    # debian: python3-requests
rules                       # debian (buster): python3-django-rules
translitcodec               # NOT IN DEBIAN
xapian-haystack             # debian (buster): python3-xapian-haystack
# xapian-haystack also requires the xapian python bindings that you’ll find
# as "python3-xapian" in Debian/Ubuntu and "python-xapian" in ArchLinux
